bugfix: Workaround bug #348068
WriteString[..] leaves an entry in Streams[] open for the
file that was just written. If that file is later deleted
(for example, by a 2nd call to NotebookWebsiteBuild[] in the
same Kernel session), the 2nd call to WriteString[..] will
silently fail, because it writes to the danging already-open
stream, and doesn't notice that the underlying file has been
deleted.

By closing the open stream, we ensure that a 2nd call to
WriteString will have to open it again, and therefore again
check the filesystem to see if the file needs to be created.