feature: Add new WrapRaised error utility function
This function is used to automatically wrap an outer (typically more general,
less technical) error around any errors raised during the evaluation of
the "wrapped" expr.

This makes it easy to attach higher-level context to exceptions raised
from lower-level code.

The first use of this function is in the main notebook build loop,
where this function is used to wrap the lower-level error about what
what wrong during the building of a notebook with a higher level
error that just says "Error building notebook" and which embeds the
name of *which* notebook was being built when the lower-level error
occurred.