bugfix: Fix incorrect escaping by switching to "XML" instead of "HTMLFragment" export format
The "HTMLFragment" export format does not escape HTML special characters
(e.g. &, <, >, etc.) that appear in textual content in the symbolic XML
expression. This makes it prohibitively difficult to ensure every place
an XMLElement is construct that we escape the HTML ourselves.

The better strategy is to use the "XML" format, which does escape things
correctly. XML`RawXML[..] can be used as an escape hatch to embed a raw
XML string into the exported string output.

The "XML" exporter requires more precisely constructed symbolic XML
constructs than the "HTMLFragment" exporter, so convertToHtml and
other functions have been adjusted to return Splice[{..}] or Nothing
instead of raw lists, to ensure that we don't construct XMLElement's
like XMLElement["..", {..}, {{..}}] (nested lists in the 3rd arg are
not allowed).

* This makes HTMLEscape[..] unused for now.