feat: Support syntax for hiding lines from HighlightSyntax cells
With this change, if a HighlightSyntax cell contains lines
with the special "begin-hidden" and "end-hidden" markers, the
enclosed set of lines will be removed from the built output.

For example, if the HighlightSyntax cell contained the following
content in the source notebook:

	# begin-hidden
	struct Foo {
		...
	}
	# end-hidden

	fn main() {
		...
	}

then the built notebook will contain just:

	fn main() {
		...
	}

Note that this also demonstrates a subtle but useful
behavior: a blank line after a hidden section a the
beginning of the cell will be hidden as well.

This ensures a visually pleasing placement of
the begin/end-hidden directives results in
a visually pleasing output as well.