feat: Add 'Preview' button to website notebook toolbar
This dropdown allows the user to quickly build and view a
preview of either the Published or Draft form of a page.

The preview HTML is built to a temporary directory and then
opened in the users default web browser.

The implementation previews the current in-memory content of
the authoring notebook (not just the latest saved changes).