fix!: Support Video assets in 'Preview'-built notebooks
This required a deeper rethinking of how the 'Preview' functionality works.
Previously, it would create a separate source directory for built notebooks.
However, that was brittle, because if a notebook depended on some other external
asset in the source build directory (like `Video[..]` objects do), that file
would not exist in the temporary source, and an error would occur when
validating `Video[..]` object results of `ConnorGray/MakeHTML` cells.

Instead of having a separate temporary notebook website source directory (which
is wasteful and complex anyway), this instead extents NotebookWebsiteBuild
with the concept of a "BuildType", which includes a new "PreviewDraftsAsPublished"
option.

This makes the 'Preview' implementation much simpler, at the expensive
of making NotebookWebsiteBuild a bit more general.

* Replace "IncludeDrafts" option with "BuildType" option, which supports the
  following values: "Published", "Drafts", "PreviewDraftsAsPublished".
* Add "FileFilterFunction" option.