feat: Support new 'Details View' sections, similar to 'Tab View' sections
These are cell groups that follow a particular structure that will
render using an HTML <details> section containing a <summary> element.

These give authors greater control over the disclousure of information,
allowing them to hide detailed information that might interrupt the
flow of the writing, but could be useful for clarification or constrast.

This finishs off and commits work from yesterday.